store_locator/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
pydantic>=2.5.3
pydantic-settings>=2.1.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2             # <--- THE FIX (Pinned Version)
python-multipart>=0.0.9
geopy>=2.4.1
redis>=5.0.1
slowapi>=0.1.9
requests>=2.31.0
pytz>=2024.1
# Testing dependencies
pytest>=8.0.0
httpx>=0.27.0